CRUD product

+++ product-api/src/main/java/com/alanjsantos/productapi/service/CategoryService.java
@@ -2,7 +2,6 @@ package com.alanjsantos.productapi.service;
 
 import com.alanjsantos.productapi.model.Category;
 import com.alanjsantos.productapi.repository.CategoryRepository;
-import com.alanjsantos.productapi.service.exception.DataIntegrityViolationException;
 import com.alanjsantos.productapi.service.exception.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,14 +19,13 @@ public class CategoryService {
         return repository.findAll();
     }
 
-    public Category getById(Integer id) {
+    public Category getById(Long id) {
         Optional<Category> obj =
-                repository.findById(id.longValue());
+                repository.findById(id);
         return obj.orElseThrow(() -> new ObjectNotFoundException("ID" + id + "This ID not aready exists in the DataBase"));
     }
 
     public Category save (Category category) {
-//        getById(category.getId());
         return repository.save(category);
 
     }
